% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{npointCheckArguments}
\alias{npointCheckArguments}
\title{Check Input Arguments}
\usage{
npointCheckArguments(args)
}
\arguments{
\item{args}{Arguments}
}
\value{
args Modified argument structure, with set files replaced with actual list of files
}
\description{
Check input arguments. At least one data set must be provided, and only
one method of parallelization should be specified
}

