% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{npointCheckSetLabels}
\alias{npointCheckSetLabels}
\title{Check Set Labels}
\usage{
npointCheckSetLabels(set, setlabels)
}
\arguments{
\item{set}{Set filename}

\item{setlabels}{Lables for items in set filename}
}
\description{
Argument checking to make sure there are labels for each data set.
}
