% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npointLmePermutation.R
\name{npointLmePermutation}
\alias{npointLmePermutation}
\title{npointLmePermutation}
\usage{
npointLmePermutation(permutationNumber, permutationRDS, targetDV,
  z_sw = TRUE, vcov = "CR2", formula, random, data,
  lmeOptsResid = list(), lmeOptsPerm = list(), lmeOpts = list(),
  sandwichOpts = list())
}
\arguments{
\item{permutationNumber}{An index specifying the row of the permutation
matrix to use. Neuropoint supplies this variable named
\code{permutationNumber} which should be passed to this argument unless the
user wishes to do something else.}

\item{permutationRDS}{The path to an RDS file that contains a permutation
matrix. This matrix should be P x N, where P is at least equal to the
number of permutations, and N is the number of rows in the design matrix.}

\item{targetDV}{A string that is the name of the dependent variable of
interest.}

\item{z_sw}{Logical indicating whether to return Z based on cluster corrected
standard errors and Satterthwaite correction for degrees of freedom.
Defaults to \code{TRUE}.}

\item{vcov}{Passed to the sandwich estimator.}

\item{formula}{A fixed-effects formula that \code{\link[nlme]{lme}}
understands.}

\item{random}{A random-effects formula that \code{\link[nlme]{lme}}
understands.}

\item{data}{The data frame appropriate for \code{\link[nlme]{lme}}.}

\item{lmeOptsResid}{Named list of any additional arguments are passed to
\code{\link[nlme]{lme}} when it's used to generate the residuals that will
be permuted.}

\item{lmeOptsPerm}{Named list of any additional arguments are passed to
\code{\link[nlme]{lme}} when it's used to estimate the full model on the
data with the permuted residuals added back in.}

\item{lmeOpts}{For backward compatibility. If this option is specified, both
lmeOptsResid and lmeOptsPerm will be set to this list.}

\item{sandwichOpts}{Named list of any additional arguments are passed to
\code{\link[clubSandwich]{coef_test}}.}
}
\value{
A list with two elements, the first of which, \code{model}, is an
  object of class "lme" representing the linear mixed-effects model fit, and
  the second of which, \code{Z}, is the Z statistic for the null hypothesis
  test for the target dependent variable named in \code{targetDV}.
}
\description{
This function permutes the data according to one row of a user supplied
permutation matrix and estimates the association between the independent
variable and the target dependent variable. This implements the Freedman-Lane
procedures (Freedman and Lane, 1983) to account for covariates properly.
}
\details{
Freedman, D., & Lane, D. (1983). A Nonstochastic Interpretation of Reported
Significance Levels. Journal of Business & Economic Statistics, 1(4),
292–298. https://doi.org/10.2307/1391660

To learn more about developing appropriate permutation matrices for complex
designs, see:

Winkler, A. M., Ridgway, G. R., Webster, M. A., Smith, S. M., & Nichols, T.
E. (2014). Permutation inference for the general linear model. NeuroImage,
92, 381–397. https://doi.org/10.1016/j.neuroimage.2014.01.060

Winkler, A. M., Webster, M. A., Vidaurre, D., Nichols, T. E., & Smith, S. M.
(2015). Multi-level block permutation. NeuroImage, 123, 253–268.
https://doi.org/10.1016/j.neuroimage.2015.05.092
}
