% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{npointMergeDesignmatWithCovariates}
\alias{npointMergeDesignmatWithCovariates}
\title{Merge the Design Matrix With Covariates}
\usage{
npointMergeDesignmatWithCovariates(desigmat, covariatefile, voxeldatdim)
}
\arguments{
\item{covariatefile}{Covariate csv file}

\item{voxeldatdim}{Dimensions of voxel data}

\item{designmat}{Design matrix}
}
\description{
Take the assembled design matrix and merge it with any covariates, making
sure that the output is reasonable.
}
