% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npointReadDataSets.R
\name{npointReadDataSets}
\alias{npointReadDataSets}
\title{Read in neuroimaging data from argument structure}
\usage{
npointReadDataSets(args, numberdatasets, mask.vertices)
}
\arguments{
\item{args}{Arguments for neuropointillist}

\item{numberdatasets}{The number of data sets specified in args}

\item{mask.vertices}{A vector of vertices used to determine what vertices to read}
}
\value{
designmat Data structure containing designmat

voxeldat Data structure containing the masked voxel data
}
\description{
This function takes an argument structure that specifies all the parameters
that should be input to npointillist and returns an array of voxel data and
a design matrix.
}

