% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{npointReadSetFiles}
\alias{npointReadSetFiles}
\title{Read Set Files}
\usage{
npointReadSetFiles(args)
}
\arguments{
\item{args}{Arguments
return args Modified argument structure, with set files replaced with actual list of files}
}
\description{
Read in the files that are in each data set and put them into the args
structure.
}
