% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{npointSplitDataSize}
\alias{npointSplitDataSize}
\title{Split Data}
\usage{
npointSplitDataSize(size, voxeldat, prefix, mask)
}
\arguments{
\item{size}{Number of voxels in each chunk}

\item{voxeldat}{Voxel data}

\item{prefix}{Text prefix to prepend to the output files}

\item{mask}{Nifti mask to split}
}
\description{
Split an input mask and data into chunks of the specified size.
}

