% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{npointWarnIfNiiFileExists}
\alias{npointWarnIfNiiFileExists}
\title{Warn if NifTI File Exists}
\usage{
npointWarnIfNiiFileExists(filename)
}
\arguments{
\item{filename}{Nifti filename}
}
\description{
Warn if a nifti file exists, and rename it if it does
}
