% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npointWriteCallingInfo.R
\name{npointWriteCallingInfo}
\alias{npointWriteCallingInfo}
\title{Write information that describes how npointillist was called}
\usage{
npointWriteCallingInfo(args)
}
\arguments{
\item{args}{Arguments provided to npointillist}
}
\description{
Write information that describes how npointillist was called
}
