% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npointWriteMakefile.R
\name{npointWriteMakefile}
\alias{npointWriteMakefile}
\title{Write an output makefile}
\usage{
npointWriteMakefile(prefix, resultnames, modelfile, designmat, makefile,
  localscript)
}
\arguments{
\item{prefix}{Prefix for output, to be prepended to outputs}

\item{resultnames}{List of names for the expected outputs}

\item{modelfile}{Name of the R model file that contains the processVoxel command}

\item{designmat}{Design matrix}

\item{makefile}{Name of makefile}

\item{localscript}{Name of script to execute makefile locally}
}
\description{
Generate a makefile for running a specific analysis
}
