% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npointWriteMakefilePermute.R
\name{npointWriteMakefilePermute}
\alias{npointWriteMakefilePermute}
\title{Write an output makefile for permutation testing}
\usage{
npointWriteMakefilePermute(prefix, resultnames, modelfile, designmat,
  makefile, localscript, npermutations)
}
\arguments{
\item{prefix}{Prefix for output, to be prepended to outputs}

\item{resultnames}{Names of results from processvoxel. You can return multiple permutation statistics, with names, and they will be used to write multiple permutation files}

\item{modelfile}{Name of the R model file that contains the processVoxel command}

\item{designmat}{Design matrix}

\item{makefile}{Name of makefile}

\item{localscript}{Name of script to execute makefile locally}

\item{npermutations}{Number of permutations to run#'}
}
\description{
Generate a makefile for running a specific analysis
}
