% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npointWriteOutputFiles.R
\name{npointWriteOutputFiles}
\alias{npointWriteOutputFiles}
\title{Write the output statistic files}
\usage{
npointWriteOutputFiles(prefix, results, mask)
}
\arguments{
\item{prefix}{Prefix for output, to be prepended to outputs}

\item{results}{Data returned from the processVoxel function}

\item{mask}{A nifti mask that corresponds to the output results}
}
\description{
This function takes an argument structure that specifies all the parameters
that should be input to npointillist and returns an array of voxel data and
a design matrix.
}

