% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npointWritePBSsubmitscript.R
\name{npointWritePBSsubmitscript}
\alias{npointWritePBSsubmitscript}
\title{Write an output PBS submit script}
\usage{
npointWritePBSsubmitscript(prefix, resultnames, modelfile, designmat,
  masterscript, jobscript, njobs, pbsPre)
}
\arguments{
\item{prefix}{Prefix for output, to be prepended to outputs}

\item{resultnames}{List of names for the expected outputs}

\item{modelfile}{Name of the model file that contains the processVoxel command}

\item{designmat}{Design matrix}

\item{masterscript}{Name of the master submit script}

\item{jobscript}{Name of the job submission script}

\item{njobs}{Number of jobs to submit}

\item{pbsPre}{The name of a preamble file for PBS - this should exist}
}
\description{
Generate an PBS submit script for the given workflow
}
\examples{

npointWritePBSsubmitscript()
}
