% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npointWriteSlurmsubmitscript.R
\name{npointWriteSlurmsubmitscript}
\alias{npointWriteSlurmsubmitscript}
\title{Write an output Slurm submit script}
\usage{
npointWriteSlurmsubmitscript(prefix, resultnames, modelfile, designmat,
  masterscript, jobscript, njobs)
}
\arguments{
\item{prefix}{Prefix for output, to be prepended to outputs}

\item{resultnames}{List of names for the expected outputs}

\item{modelfile}{Name of the model file that contains the processVoxel command}

\item{designmat}{Design matrix}

\item{masterscript}{Name of the master submit script}

\item{jobscript}{Name of the job submission script}

\item{njobs}{Number of jobs to submit}
}
\description{
Generate an Slurm submit script for the given workflow
}
\examples{

npointWriteSlurmsubmitscript()
}
